HotApp / WarmApp and EstablishedApp distinction
The governor needs to take care about three types of protocols: ones
that run for warm peer and hot peers.  For this purpose we provide
a distinction into three disjoint subsets: `WarmApp` warm protocols,
which run for  warm peers, `HotApp` for hot peers and `EstablishedApp`
for protocols which must run in for either hot or warm peers
(established peers in the language for p2p-governor).  Using GADTs we
make this distinction at type level.  This patch provides convientient
functions to work with the introduced complexity.